Criei o Menu de Chacara. Alterei áreaTotal para Double.

+++ RP2T3/src/Imovel/Imovel.java
@@ -138,7 +138,7 @@ public abstract class Imovel {
     /**
      * @param areaTotal the areaTotal to set
      */
-    public void setAreaTotal(int areaTotal) {
+    public void setAreaTotal(double areaTotal) {
         this.areaTotal = areaTotal;
     }
